Renamed views to screens to prevent confusing between my "Screens/pages" and for example ListView
