+++ lib/screens/home_screen.dart
@@ -3,7 +3,7 @@ import 'package:provider/provider.dart';
 import 'package:space_x_tracker/custom_color_scheme.dart';
 import 'package:space_x_tracker/providers/launch_provider.dart';
 import 'package:space_x_tracker/providers/models/launch.dart';
-import 'package:space_x_tracker/views/filter_view.dart';
+import 'package:space_x_tracker/screens/filter_screen.dart';
 import 'package:space_x_tracker/widgets/home/card_list_view.dart';
 import 'package:http/http.dart' as http;
 import 'package:space_x_tracker/widgets/flash_message.dart';
