+++ test/widget/screens/home_view_test.dart
@@ -11,8 +11,8 @@ import 'package:provider/provider.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:space_x_tracker/project_theme.dart';
 import 'package:space_x_tracker/providers/launch_provider.dart';
-import 'package:space_x_tracker/views/filter_view.dart';
-import 'package:space_x_tracker/views/home_view.dart';
+import 'package:space_x_tracker/screens/filter_screen.dart';
+import 'package:space_x_tracker/screens/home_screen.dart';
 import 'package:space_x_tracker/widgets/home/launch_card.dart';
 
 import 'home_view_test.mocks.dart';
