+++ test/widget/screens/rocket_view_test.dart
@@ -12,7 +12,7 @@ import 'package:space_x_tracker/project_theme.dart';
 import 'package:space_x_tracker/providers/models/launch.dart';
 import 'package:space_x_tracker/providers/models/rocket.dart';
 import 'package:space_x_tracker/providers/rocket_provider.dart';
-import 'package:space_x_tracker/views/rocket_view.dart';
+import 'package:space_x_tracker/screens/rocket_screen.dart';
 import 'package:space_x_tracker/widgets/home/launch_card.dart';
 import 'package:space_x_tracker/widgets/rocket/rocket_cover_image.dart';
